+++ .gitignore
@@ -126,6 +126,9 @@ _ReSharper*/
 *.[Rr]e[Ss]harper
 *.DotSettings.user
 
+# Rider
+.idea/
+
 # JustCode is a .NET coding add-in
 .JustCode
