+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultCollections/ResultCollectionBindTryWhereAsyncExtensions.cs
@@ -5,31 +5,43 @@ using System.Threading.Tasks;
 using ResultFunctional.Models.Implementations.Results;
 using ResultFunctional.Models.Interfaces.Errors.Base;
 using ResultFunctional.Models.Interfaces.Results;
-using static ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultCollections.ResultCollectionBindTryAsyncExtensions;
+using static ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultCollections.ResultCollectionBindTryTaskAsyncExtensions;
 
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultCollections
 {
     /// <summary>
-    /// Методы расширения для результирующего ответа со значением и обработкой исключений асинхронно
+    /// Extension methods for result collection monad async function with conditions and exception handling
     /// </summary>
     public static class ResultCollectionBindTryWhereAsyncExtensions
     {
         /// <summary>
-        /// Результирующий ответ со связыванием со значением с обработкой функции при положительном условии для задачи-объекта
+        /// Execute result collection async function in no error case; else catch exception
         /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="func">Monad result collection function</param>
+        /// <param name="exceptionFunc">Exception function</param>
+        /// <returns>Outgoing result collection</returns>
         public static async Task<IResultCollection<TValueOut>> ResultCollectionBindTryOkAsync<TValueIn, TValueOut>(this IResultCollection<TValueIn> @this,
                                                                                                        Func<IEnumerable<TValueIn>, Task<IResultCollection<TValueOut>>> func,
                                                                                                        Func<Exception, IErrorResult> exceptionFunc) =>
             await @this.
-            ResultCollectionBindOkAsync(value => ResultCollectionBindTryAsync(() => func.Invoke(value), exceptionFunc));
+            ResultCollectionBindOkAsync(value => ResultCollectionBindTryTaskAsync(() => func.Invoke(value), exceptionFunc));
 
         /// <summary>
-        /// Результирующий ответ со связыванием со значением с обработкой функции при положительном условии для задачи-объекта
+        /// Execute result collection async function in no error case; else catch exception
         /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="func">Monad result collection function</param>
+        /// <param name="error">Error</param>
+        /// <returns>Outgoing result collection</returns>
         public static async Task<IResultCollection<TValueOut>> ResultCollectionBindTryOkAsync<TValueIn, TValueOut>(this IResultCollection<TValueIn> @this,
                                                                                                        Func<IEnumerable<TValueIn>, Task<IResultCollection<TValueOut>>> func,
                                                                                                        IErrorResult error) =>
             await @this.
-            ResultCollectionBindOkAsync(value => ResultCollectionBindTryAsync(() => func.Invoke(value), error));
+            ResultCollectionBindOkAsync(value => ResultCollectionBindTryTaskAsync(() => func.Invoke(value), error));
     }
 }
