+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultCollections/ResultCollectionBindWhereAsyncExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using ResultFunctional.FunctionalExtensions.Sync.ResultExtension.ResultCollections;
 using ResultFunctional.FunctionalExtensions.Sync.ResultExtension.ResultErrors;
 using ResultFunctional.Models.Implementations.Results;
 using ResultFunctional.Models.Interfaces.Errors.Base;
@@ -9,64 +10,104 @@ using ResultFunctional.Models.Interfaces.Results;
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultCollections
 {
     /// <summary>
-    /// Обработка условий для результирующего асинхронного связывающего ответа с коллекцией
+    /// Extension methods for result collection monad async function with conditions
     /// </summary>
     public static class ResultCollectionBindWhereAsyncExtensions
     {
         /// <summary>
-        /// Выполнение условия или возвращение предыдущей ошибки в асинхронном результирующем ответе с коллекцией
-        /// </summary>      
+        /// Execute monad result collection async function base on predicate condition
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function returning errors if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result collection</returns>
         public static async Task<IResultCollection<TValueOut>> ResultCollectionBindContinueAsync<TValueIn, TValueOut>(this IResultCollection<TValueIn> @this,
-                                                                                                            Func<IReadOnlyCollection<TValueIn>, bool> predicate,
-                                                                                                            Func<IReadOnlyCollection<TValueIn>, Task<IResultCollection<TValueOut>>> okFunc,
-                                                                                                            Func<IReadOnlyCollection<TValueIn>, Task<IEnumerable<IErrorResult>>> badFunc) =>
-         @this.OkStatus
-             ? predicate(@this.Value)
-                 ? await okFunc.Invoke(@this.Value)
-                 : new ResultCollection<TValueOut>(await badFunc.Invoke(@this.Value))
-             : new ResultCollection<TValueOut>(@this.Errors);
+                                                                                                                      Func<IReadOnlyCollection<TValueIn>, bool> predicate,
+                                                                                                                      Func<IReadOnlyCollection<TValueIn>, Task<IResultCollection<TValueOut>>> okFunc,
+                                                                                                                      Func<IReadOnlyCollection<TValueIn>, Task<IEnumerable<IErrorResult>>> badFunc) =>
+            @this.OkStatus
+                ? predicate(@this.Value)
+                    ? await okFunc.Invoke(@this.Value)
+                    : new ResultCollection<TValueOut>(await badFunc.Invoke(@this.Value))
+                : new ResultCollection<TValueOut>(@this.Errors);
 
         /// <summary>
-        /// Выполнение условия или возвращение предыдущей ошибки в асинхронном результирующем ответе с коллекцией
-        /// </summary>      
+        /// Execute monad result collection async function base on predicate condition
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result collection</returns>
         public static async Task<IResultCollection<TValueOut>> ResultCollectionBindWhereAsync<TValueIn, TValueOut>(this IResultCollection<TValueIn> @this,
-                                                                                                            Func<IReadOnlyCollection<TValueIn>, bool> predicate,
-                                                                                                            Func<IReadOnlyCollection<TValueIn>, Task<IResultCollection<TValueOut>>> okFunc,
-                                                                                                            Func<IReadOnlyCollection<TValueIn>, Task<IResultCollection<TValueOut>>> badFunc) =>
-         @this.OkStatus
-             ? predicate(@this.Value)
-                 ? await okFunc.Invoke(@this.Value)
-                 : await badFunc.Invoke(@this.Value)
-             : new ResultCollection<TValueOut>(@this.Errors);
+                                                                                                                   Func<IReadOnlyCollection<TValueIn>, bool> predicate,
+                                                                                                                   Func<IReadOnlyCollection<TValueIn>, Task<IResultCollection<TValueOut>>> okFunc,
+                                                                                                                   Func<IReadOnlyCollection<TValueIn>, Task<IResultCollection<TValueOut>>> badFunc) =>
+            @this.OkStatus
+                ? predicate(@this.Value)
+                    ? await okFunc.Invoke(@this.Value)
+                    : await badFunc.Invoke(@this.Value)
+                : new ResultCollection<TValueOut>(@this.Errors);
 
         /// <summary>
-        /// Выполнение положительного условия результирующего асинхронного ответа со связыванием или возвращение предыдущей ошибки в результирующем ответе с коллекцией
-        /// </summary>   
+        /// Execute monad result collection async function depending on result collection errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="okFunc">Function if result collection hasn't errors</param>
+        /// <param name="badFunc">Function if result collection has errors</param>
+        /// <returns>Outgoing result collection</returns>
+        public static async Task<IResultCollection<TValueOut>> ResultCollectionBindOkBadAsync<TValueIn, TValueOut>(this IResultCollection<TValueIn> @this,
+                                                                                                                   Func<IReadOnlyCollection<TValueIn>, Task<IResultCollection<TValueOut>>> okFunc,
+                                                                                                                   Func<IReadOnlyCollection<IErrorResult>, Task<IResultCollection<TValueOut>>> badFunc) =>
+            @this.OkStatus
+                ? await okFunc.Invoke(@this.Value)
+                : await badFunc.Invoke(@this.Errors);
+
+        /// <summary>
+        /// Execute monad result collection async function if incoming result collection hasn't errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="okFunc">Function if incoming result collection hasn't errors</param>
+        /// <returns>Outgoing result collection</returns>
         public static async Task<IResultCollection<TValueOut>> ResultCollectionBindOkAsync<TValueIn, TValueOut>(this IResultCollection<TValueIn> @this,
-                                                                                                      Func<IReadOnlyCollection<TValueIn>, Task<IResultCollection<TValueOut>>> okFunc) =>
+                                                                                                                Func<IReadOnlyCollection<TValueIn>, Task<IResultCollection<TValueOut>>> okFunc) =>
             @this.OkStatus
                 ? await okFunc.Invoke(@this.Value)
                 : new ResultCollection<TValueOut>(@this.Errors);
 
-
         /// <summary>
-        /// Выполнение негативного условия результирующего асинхронного ответа или возвращение положительного в результирующем ответе с коллекцией
-        /// </summary>   
+        /// Execute monad result collection async function if incoming result collection has errors
+        /// </summary>
+        /// <typeparam name="TValue">Result type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="badFunc">Function if incoming result collection has errors</param>
+        /// <returns>Outgoing result collection</returns>
         public static async Task<IResultCollection<TValue>> ResultCollectionBindBadAsync<TValue>(this IResultCollection<TValue> @this,
-                                                                                       Func<IReadOnlyCollection<IErrorResult>, Task<IResultCollection<TValue>>> badFunc) =>
+                                                                                                 Func<IReadOnlyCollection<IErrorResult>, Task<IResultCollection<TValue>>> badFunc) =>
             @this.OkStatus
                 ? @this
                 : await badFunc.Invoke(@this.Errors);
 
         /// <summary>
-        /// Добавить асинхронно ошибки результирующего ответа или вернуть результат с ошибками для ответа с коллекцией
+        /// Adding errors async to result collection if ones hasn't errors
         /// </summary>
+        /// <typeparam name="TValue">Result type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="okFunc">Error function if incoming result collection hasn't errors</param>
+        /// <returns>Outgoing result collection</returns>
         public static async Task<IResultCollection<TValue>> ResultCollectionBindErrorsOkAsync<TValue>(this IResultCollection<TValue> @this,
-                                                                                       Func<IReadOnlyCollection<TValue>, Task<IResultError>> okFunc) =>
+                                                                                                      Func<IReadOnlyCollection<TValue>, Task<IResultError>> okFunc) =>
             await @this.
-            ResultCollectionBindOkAsync(collection => okFunc.Invoke(collection).
-                                                      MapTaskAsync(resultError => resultError.ToResultCollection(collection)));
-
-
+                ResultCollectionBindOkAsync(collection => okFunc.Invoke(collection).
+                                                                 MapTaskAsync(resultError => resultError.ToResultCollection(collection)));
     }
 }
