+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultCollections/ResultCollectionBindWhereBindAsyncExtensions.cs
@@ -7,13 +7,20 @@ using ResultFunctional.Models.Interfaces.Results;
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultCollections
 {
     /// <summary>
-    /// Обработка условий для асинхронного результирующего асинхронного связывающего ответа с коллекцией для задачи-объекта
+    /// Extension methods for result collection monad async task function with conditions
     /// </summary>
     public static class ResultCollectionBindWhereBindAsyncExtensions
     {
         /// <summary>
-        /// Выполнение условия или возвращение предыдущей ошибки в результирующем ответе с коллекцией задачи-объекта
-        /// </summary>      
+        /// Execute monad result collection async task function base on predicate condition
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function returning errors if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result collection</returns>
         public static async Task<IResultCollection<TValueOut>> ResultCollectionBindContinueBindAsync<TValueIn, TValueOut>(this Task<IResultCollection<TValueIn>> @this,
                                                                                                             Func<IReadOnlyCollection<TValueIn>, bool> predicate,
                                                                                                             Func<IReadOnlyCollection<TValueIn>, Task<IResultCollection<TValueOut>>> okFunc,
@@ -22,8 +29,15 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultColl
             MapBindAsync(awaitedThis => awaitedThis.ResultCollectionBindContinueAsync(predicate, okFunc, badFunc));
 
         /// <summary>
-        /// Выполнение условия или возвращение предыдущей ошибки в результирующем ответе с коллекцией задачи-объекта
-        /// </summary>      
+        /// Execute monad result collection async task function base on predicate condition
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result collection</returns>
         public static async Task<IResultCollection<TValueOut>> ResultCollectionBindWhereBindAsync<TValueIn, TValueOut>(this Task<IResultCollection<TValueIn>> @this,
                                                                                                             Func<IReadOnlyCollection<TValueIn>, bool> predicate,
                                                                                                             Func<IReadOnlyCollection<TValueIn>, Task<IResultCollection<TValueOut>>> okFunc,
@@ -32,24 +46,52 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultColl
             MapBindAsync(awaitedThis => awaitedThis.ResultCollectionBindWhereAsync(predicate, okFunc, badFunc));
 
         /// <summary>
-        /// Выполнение положительного условия асинхронного результирующего ответа или возвращение предыдущей ошибки в результирующем ответе с коллекцией для задачи-объекта
-        /// </summary>   
+        /// Execute monad result collection async task function depending on result collection errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="okFunc">Function if result collection hasn't errors</param>
+        /// <param name="badFunc">Function if result collection has errors</param>
+        /// <returns>Outgoing result collection</returns>
+        public static async Task<IResultCollection<TValueOut>> ResultCollectionBindOkBadBindAsync<TValueIn, TValueOut>(this Task<IResultCollection<TValueIn>> @this,
+                                                                                                                   Func<IReadOnlyCollection<TValueIn>, Task<IResultCollection<TValueOut>>> okFunc,
+                                                                                                                   Func<IReadOnlyCollection<IErrorResult>, Task<IResultCollection<TValueOut>>> badFunc) =>
+            await @this.
+            MapBindAsync(awaitedThis => awaitedThis.ResultCollectionBindOkBadAsync(okFunc, badFunc));
+
+        /// <summary>
+        /// Execute monad result collection async task function if incoming result collection hasn't errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="okFunc">Function if incoming result collection hasn't errors</param>
+        /// <returns>Outgoing result collection</returns>
         public static async Task<IResultCollection<TValueOut>> ResultCollectionBindOkBindAsync<TValueIn, TValueOut>(this Task<IResultCollection<TValueIn>> @this,
                                                                                                       Func<IReadOnlyCollection<TValueIn>, Task<IResultCollection<TValueOut>>> okFunc) =>
             await @this.
             MapBindAsync(awaitedThis => awaitedThis.ResultCollectionBindOkAsync(okFunc));
 
         /// <summary>
-        /// Выполнение негативного условия асинхронного результирующего ответа или возвращение положительного в результирующем ответе с коллекцией для задачи-объекта
-        /// </summary>   
+        /// Execute monad result collection async task function if incoming result collection has errors
+        /// </summary>
+        /// <typeparam name="TValue">Result type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="badFunc">Function if incoming result collection has errors</param>
+        /// <returns>Outgoing result collection</returns>
         public static async Task<IResultCollection<TValue>> ResultCollectionBindBadBindAsync<TValue>(this Task<IResultCollection<TValue>> @this,
                                                                                        Func<IReadOnlyCollection<IErrorResult>, Task<IResultCollection<TValue>>> badFunc) =>
             await @this.
             MapBindAsync(awaitedThis => awaitedThis.ResultCollectionBindBadAsync(badFunc));
 
         /// <summary>
-        /// Добавить ошибки асинхронного результирующего ответа или вернуть результат с ошибками для ответа с коллекцией для задачи-объекта
+        /// Adding errors async to task result collection if ones hasn't errors
         /// </summary>
+        /// <typeparam name="TValue">Result type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="okFunc">Error function if incoming result collection hasn't errors</param>
+        /// <returns>Outgoing result collection</returns>
         public static async Task<IResultCollection<TValue>> ResultCollectionBindErrorsOkBindAsync<TValue>(this Task<IResultCollection<TValue>> @this,
                                                                                        Func<IReadOnlyCollection<TValue>, Task<IResultError>> okFunc) =>
             await @this.
