+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultCollections/ResultCollectionBindWhereTaskAsyncExtensions.cs
@@ -8,13 +8,20 @@ using ResultFunctional.Models.Interfaces.Results;
 namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultCollections
 {
     /// <summary>
-    /// Обработка условий для результирующего асинхронного связывающего ответа с коллекцией для задачи-объекта
+    /// Extension methods for result collection monad task function with conditions
     /// </summary>
     public static class ResultCollectionBindWhereTaskAsyncExtensions
     {
         /// <summary>
-        /// Выполнение условия или возвращение предыдущей ошибки в результирующем ответе с коллекцией задачи-объекта
-        /// </summary>      
+        /// Execute monad result collection task function base on predicate condition
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function returning errors if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result collection</returns> 
         public static async Task<IResultCollection<TValueOut>> ResultCollectionBindContinueTaskAsync<TValueIn, TValueOut>(this Task<IResultCollection<TValueIn>> @this,
                                                                                                             Func<IReadOnlyCollection<TValueIn>, bool> predicate,
                                                                                                             Func<IReadOnlyCollection<TValueIn>, IResultCollection<TValueOut>> okFunc,
@@ -22,40 +29,72 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultColl
             await @this.
             MapTaskAsync(awaitedThis => awaitedThis.ResultCollectionBindContinue(predicate, okFunc, badFunc));
 
-        /// <summary>
-        /// Выполнение условия или возвращение предыдущей ошибки в результирующем ответе с коллекцией задачи-объекта
-        /// </summary>      
+       /// <summary>
+        /// Execute monad result collection task function base on predicate condition
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function if predicate <see langword="false"/></param>
+        /// <returns>Outgoing result collection</returns> 
         public static async Task<IResultCollection<TValueOut>> ResultCollectionBindWhereTaskAsync<TValueIn, TValueOut>(this Task<IResultCollection<TValueIn>> @this,
                                                                                                             Func<IReadOnlyCollection<TValueIn>, bool> predicate,
                                                                                                             Func<IReadOnlyCollection<TValueIn>, IResultCollection<TValueOut>> okFunc,
                                                                                                             Func<IReadOnlyCollection<TValueIn>, IResultCollection<TValueOut>> badFunc) =>
             await @this.
             MapTaskAsync(awaitedThis => awaitedThis.ResultCollectionBindWhere(predicate, okFunc, badFunc));
+       
+        /// <summary>
+        /// Execute monad result collection task function depending on result collection errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="okFunc">Function if result collection hasn't errors</param>
+        /// <param name="badFunc">Function if result collection has errors</param>
+        /// <returns>Outgoing result collection</returns>
+        public static async Task<IResultCollection<TValueOut>> ResultCollectionBindOkBadTaskAsync<TValueIn, TValueOut>(this Task<IResultCollection<TValueIn>> @this,
+                                                                                        Func<IReadOnlyCollection<TValueIn>, IResultCollection<TValueOut>> okFunc,
+                                                                                        Func<IReadOnlyCollection<IErrorResult>, IResultCollection<TValueOut>> badFunc) =>
+            await @this.
+            MapTaskAsync(awaitedThis => awaitedThis.ResultCollectionBindOkBad(okFunc, badFunc));
 
         /// <summary>
-        /// Выполнение положительного условия результирующего ответа или возвращение предыдущей ошибки в результирующем ответе с коллекцией для задачи-объекта
-        /// </summary>   
-        public static async Task<IResultCollection<TValueOut>> ResultCollectionBindOkTaskAsync<TValueIn, TValueOut>(this Task<IResultCollection<TValueIn>> @this,
-                                                                                                      Func<IReadOnlyCollection<TValueIn>, IResultCollection<TValueOut>> okFunc) =>
+        /// Execute monad result collection task function if incoming result collection hasn't errors
+        /// </summary>
+        /// <typeparam name="TValueIn">Incoming type</typeparam>
+        /// <typeparam name="TValueOut">Outgoing type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="okFunc">Function if incoming result collection hasn't errors</param>
+        /// <returns>Outgoing result collection</returns>
+        public static async Task<IResultCollection<TValueOut>> ResultCollectionBindOkTaskAsync<TValueIn, TValueOut>(this Task<IResultCollection<TValueIn>> @this, Func<IReadOnlyCollection<TValueIn>, IResultCollection<TValueOut>> okFunc) =>
             await @this.
             MapTaskAsync(awaitedThis => awaitedThis.ResultCollectionBindOk(okFunc));
 
         /// <summary>
-        /// Выполнение негативного условия результирующего ответа или возвращение положительного в результирующем ответе с коллекцией для задачи-объекта
-        /// </summary>   
+        /// Execute monad result collection task function if incoming result collection has errors
+        /// </summary>
+        /// <typeparam name="TValue">Result type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="badFunc">Function if incoming result collection has errors</param>
+        /// <returns>Outgoing result collection</returns> 
         public static async Task<IResultCollection<TValue>> ResultCollectionBindBadTaskAsync<TValue>(this Task<IResultCollection<TValue>> @this,
                                                                                        Func<IReadOnlyCollection<IErrorResult>, IResultCollection<TValue>> badFunc) =>
             await @this.
             MapTaskAsync(awaitedThis => awaitedThis.ResultCollectionBindBad(badFunc));
 
         /// <summary>
-        /// Добавить ошибки результирующего ответа или вернуть результат с ошибками для ответа с коллекцией для задачи-объекта
+        /// Adding errors to task result collection if ones hasn't errors
         /// </summary>
+        /// <typeparam name="TValue">Result type</typeparam>
+        /// <param name="this">Incoming result collection</param>
+        /// <param name="okFunc">Error function if incoming result collection hasn't errors</param>
+        /// <returns>Outgoing result collection</returns>
         public static async Task<IResultCollection<TValue>> ResultCollectionBindErrorsOkTaskAsync<TValue>(this Task<IResultCollection<TValue>> @this,
                                                                                        Func<IReadOnlyCollection<TValue>, IResultError> okFunc) =>
             await @this.
             MapTaskAsync(awaitedThis => awaitedThis.ResultCollectionBindErrorsOk(okFunc));
-
-
     }
 }
