+++ ResultFunctional/FunctionalExtensions/Async/ResultExtension/ResultValues/ResultValueTaskAsyncExtensions.cs
@@ -14,55 +14,6 @@ namespace ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValu
     public static class ResultValueTaskAsyncExtensions
     {
         /// <summary>
-        /// Преобразовать в результирующий ответ со значением в коллекцию задачи-объекта
-        /// </summary>      
-        public static async Task<IResultCollection<TValue>> ToResultCollectionTaskAsync<TValue>(this Task<IEnumerable<IResultValue<TValue>>> @this) =>
-            await @this.
-            MapTaskAsync(awaitedThis => awaitedThis.ToResultCollection());
-
-        /// <summary>
-        /// Преобразовать в результирующий ответ со значением в коллекцию задачи-объекта
-        /// </summary>      
-        public static async Task<IResultCollection<TValue>> ToResultCollectionTaskAsync<TValue>(this Task<IList<IResultValue<TValue>>> @this) =>
-            await @this.
-            MapTaskAsync(awaitedThis => awaitedThis.ToResultCollection());
-
-        /// <summary>
-        /// Преобразовать в результирующий ответ со значением в коллекцию задачи-объекта
-        /// </summary>      
-        public static async Task<IResultCollection<TValue>> ToResultCollectionTaskAsync<TValue>(this Task<IReadOnlyCollection<IResultValue<TValue>>> @this) =>
-            await @this.
-            MapTaskAsync(awaitedThis => awaitedThis.ToResultCollection());
-
-        /// <summary>
-        /// Преобразовать в результирующий ответ со значением в коллекцию задачи-объекта
-        /// </summary>      
-        public static async Task<IResultCollection<TValue>> ToResultCollectionTaskAsync<TValue>(this Task<IResultValue<IEnumerable<TValue>>> @this) =>
-            await @this.
-            MapTaskAsync(awaitedThis => awaitedThis.ToResultCollection());
-
-        /// <summary>
-        /// Преобразовать в результирующий ответ со значением в коллекцию задачи-объекта
-        /// </summary>      
-        public static async Task<IResultCollection<TValue>> ToResultCollectionTaskAsync<TValue>(this Task<IResultValue<IReadOnlyCollection<TValue>>> @this) =>
-            await @this.
-            MapTaskAsync(awaitedThis => awaitedThis.ToResultCollection());
-
-        /// <summary>
-        /// Преобразовать в результирующий ответ со значением в коллекцию задачи-объекта
-        /// </summary>      
-        public static async Task<IResultCollection<TValue>> ToResultCollectionTaskAsync<TValue>(this Task<IResultValue<ReadOnlyCollection<TValue>>> @this) =>
-            await @this.
-            MapTaskAsync(awaitedThis => awaitedThis.ToResultCollection());
-
-        /// <summary>
-        /// Преобразовать в результирующий ответ со значением в коллекцию задачи-объекта
-        /// </summary>      
-        public static async Task<IResultCollection<TValue>> ToResultCollectionTaskAsync<TValue>(this Task<IResultValue<List<TValue>>> @this) =>
-            await @this.
-            MapTaskAsync(awaitedThis => awaitedThis.ToResultCollection());
-
-        /// <summary>
         /// Преобразовать в результирующий ответ
         /// </summary>      
         public static async Task<IResultError> ToResultErrorTaskAsync<TValue>(this Task<IResultValue<TValue>> @this) =>
