+++ ResultFunctional/FunctionalExtensions/Async/WhereAsyncExtensions.cs
@@ -4,13 +4,20 @@ using System.Threading.Tasks;
 namespace ResultFunctional.FunctionalExtensions.Async
 {
     /// <summary>
-    /// Методы расширения для проверки условий асинхронно
+    /// Async extension methods for condition functions
     /// </summary>
     public static class WhereAsyncExtensions
     {
         /// <summary>
-        /// Условие продолжающее действие
-        /// </summary>      
+        /// Execute converting async function base on predicate condition
+        /// </summary>
+        /// <typeparam name="TSource">Source type</typeparam>
+        /// <typeparam name="TResult">Result type</typeparam>
+        /// <param name="this">Source</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function if predicate <see langword="false"/></param>
+        /// <returns>Converting function result</returns>
         public static async Task<TResult> WhereContinueAsync<TSource, TResult>(this TSource @this,
                                                                               Func<TSource, bool> predicate,
                                                                               Func<TSource, Task<TResult>> okFunc,
@@ -20,16 +27,26 @@ namespace ResultFunctional.FunctionalExtensions.Async
                 : await badFunc.Invoke(@this);
 
         /// <summary>
-        /// Обработка позитивного условия
-        /// </summary>      
+        /// Execute converting async function if predicate condition <see langword="true"/>
+        /// </summary>
+        /// <typeparam name="TSource">Source type</typeparam>
+        /// <param name="this">Source</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <returns>Converting function result</returns>  
         public static async Task<TSource> WhereOkAsync<TSource>(this TSource @this,
                                                                Func<TSource, bool> predicate,
                                                                Func<TSource, Task<TSource>> okFunc) =>
              await @this.WhereContinueAsync(predicate, okFunc, _ => Task.FromResult(@this));
 
         /// <summary>
-        /// Обработка негативного условия
-        /// </summary>      
+        /// Execute converting async function if predicate condition <see langword="false"/>
+        /// </summary>
+        /// <typeparam name="TSource">Source type</typeparam>
+        /// <param name="this">Source</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="badFunc">Function if predicate <see langword="false"/></param>
+        /// <returns>Converting function result</returns>
         public static async Task<TSource> WhereBadAsync<TSource>(this TSource @this,
                                                                 Func<TSource, bool> predicate,
                                                                 Func<TSource, Task<TSource>> badFunc) =>
