+++ ResultFunctional/FunctionalExtensions/Async/WhereBindAsyncExtensions.cs
@@ -4,13 +4,20 @@ using System.Threading.Tasks;
 namespace ResultFunctional.FunctionalExtensions.Async
 {
     /// <summary>
-    /// Методы расширения для проверки условий асинхронно для задачи-объекта
+    /// Async extension methods for task condition functions
     /// </summary>
     public static class WhereBindAsyncExtensions
     {
         /// <summary>
-        /// Условие продолжающее действие
-        /// </summary>      
+        /// Execute converting async task function base on predicate condition
+        /// </summary>
+        /// <typeparam name="TSource">Source type</typeparam>
+        /// <typeparam name="TResult">Result type</typeparam>
+        /// <param name="this">Source</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <param name="badFunc">Function if predicate <see langword="false"/></param>
+        /// <returns>Converting function result</returns>      
         public static async Task<TResult> WhereContinueBindAsync<TSource, TResult>(this Task<TSource> @this,
                                                                                    Func<TSource, bool> predicate,
                                                                                    Func<TSource, Task<TResult>> okFunc,
@@ -19,8 +26,13 @@ namespace ResultFunctional.FunctionalExtensions.Async
             MapBindAsync(thisAwaited => thisAwaited.WhereContinueAsync(predicate, okFunc, badFunc));
 
         /// <summary>
-        /// Обработка позитивного условия
-        /// </summary>      
+        /// Execute converting async task function if predicate condition <see langword="true"/>
+        /// </summary>
+        /// <typeparam name="TSource">Source type</typeparam>
+        /// <param name="this">Source</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="okFunc">Function if predicate <see langword="true"/></param>
+        /// <returns>Converting function result</returns>         
         public static async Task<TSource> WhereOkBindAsync<TSource>(this Task<TSource> @this,
                                                                   Func<TSource, bool> predicate,
                                                                   Func<TSource, Task<TSource>> okFunc) =>
@@ -28,8 +40,13 @@ namespace ResultFunctional.FunctionalExtensions.Async
             MapBindAsync(thisAwaited => thisAwaited.WhereOkAsync(predicate, okFunc));
 
         /// <summary>
-        /// Обработка негативного условия
-        /// </summary>      
+        /// Execute converting async task function if predicate condition <see langword="false"/>
+        /// </summary>
+        /// <typeparam name="TSource">Source type</typeparam>
+        /// <param name="this">Source</param>
+        /// <param name="predicate">Predicate function</param>
+        /// <param name="badFunc">Function if predicate <see langword="false"/></param>
+        /// <returns>Converting function result</returns>     
         public static async Task<TSource> WhereBadBindAsync<TSource>(this Task<TSource> @this,
                                                                      Func<TSource, bool> predicate,
                                                                      Func<TSource, Task<TSource>> badFunc) =>
