+++ ResultFunctional/ResultFunctional.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <Product>Functional Library ($(TargetFramework))</Product>
-    <TargetFrameworks>net6.0;net5.0;netstandard2.0;netstandard2.1;net461</TargetFrameworks>
+    <TargetFrameworks>net6.0;netstandard2.0;netstandard2.1;net461</TargetFrameworks>
     <Nullable>enable</Nullable>
     <Authors>Ivan Rubilnik</Authors>
     <Company />
