+++ ResultFunctionalXUnit/FunctionalExtensions/Async/ResultExtension/ResultValueTest/ResultValueTaskAsyncExtensionsTest.cs
@@ -2,6 +2,7 @@
 using System.Collections.ObjectModel;
 using System.Linq;
 using System.Threading.Tasks;
+using ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultCollections;
 using ResultFunctional.FunctionalExtensions.Async.ResultExtension.ResultValues;
 using ResultFunctional.FunctionalExtensions.Sync.ResultExtension.ResultValues;
 using ResultFunctional.Models.Implementations.ResultFactory;
@@ -17,131 +18,6 @@ namespace ResultFunctionalXUnit.FunctionalExtensions.Async.ResultExtension.Resul
     /// </summary>
     public class ResultValueTaskAsyncExtensionsTest
     {
-        /// <summary>
-        /// Вернуть результирующий ответ задачи-объекта с коллекцией без ошибок
-        /// </summary>      
-        [Fact]
-        public async Task ToResultCollectionTaskAsync_Enumerable_OkStatus()
-        {
-            var collection = GetRangeNumber();
-            var resultNoError = ResultValueFactory.CreateTaskResultValue<IEnumerable<int>>(collection);
-
-            var resultValue = await resultNoError.ToResultCollectionTaskAsync();
-
-            Assert.True(resultValue.OkStatus);
-            Assert.True(collection.SequenceEqual(resultValue.Value));
-        }
-
-        /// <summary>
-        /// Вернуть результирующий ответ задачи-объекта со значением с ошибкой
-        /// </summary>      
-        [Fact]
-        public async Task ToResultCollectionTaskAsync_Enumerable_HasErrors()
-        {
-            var error = CreateErrorTest();
-            var resultHasError = ResultValueFactory.CreateTaskResultValueError<IEnumerable<int>>(error);
-
-            var resultValue = await resultHasError.ToResultCollectionTaskAsync();
-
-            Assert.True(resultValue.HasErrors);
-            Assert.Single(resultValue.Errors);
-            Assert.True(error.Equals(resultValue.Errors.Last()));
-        }
-
-        /// <summary>
-        /// Вернуть результирующий ответ задачи-объекта с коллекцией без ошибок
-        /// </summary>      
-        [Fact]
-        public async Task ToResultCollectionTaskAsync_IReadOnlyCollection_OkStatus()
-        {
-            var collection = GetRangeNumber();
-            var resultNoError = ResultValueFactory.CreateTaskResultValue(collection);
-
-            var resultValue = await resultNoError.ToResultCollectionTaskAsync();
-
-            Assert.True(resultValue.OkStatus);
-            Assert.True(collection.SequenceEqual(resultValue.Value));
-        }
-
-        /// <summary>
-        /// Вернуть результирующий ответ задачи-объекта со значением с ошибкой
-        /// </summary>      
-        [Fact]
-        public async Task ToResultCollectionTaskAsync_IReadOnlyCollection_HasErrors()
-        {
-            var error = CreateErrorTest();
-            var resultHasError = ResultValueFactory.CreateTaskResultValueError<IReadOnlyCollection<int>>(error);
-
-            var resultValue = await resultHasError.ToResultCollectionTaskAsync();
-
-            Assert.True(resultValue.HasErrors);
-            Assert.Single(resultValue.Errors);
-            Assert.True(error.Equals(resultValue.Errors.Last()));
-        }
-
-        /// <summary>
-        /// Вернуть результирующий ответ задачи-объекта с коллекцией без ошибок
-        /// </summary>      
-        [Fact]
-        public async Task ToResultCollectionTaskAsync_ReadOnlyCollection_OkStatus()
-        {
-            var collection = GetRangeNumber().ToList().AsReadOnly();
-            var resultNoError = ResultValueFactory.CreateTaskResultValue(collection);
-
-            var resultValue = await resultNoError.ToResultCollectionTaskAsync();
-
-            Assert.True(resultValue.OkStatus);
-            Assert.True(collection.SequenceEqual(resultValue.Value));
-        }
-
-        /// <summary>
-        /// Вернуть результирующий ответ задачи-объекта со значением с ошибкой
-        /// </summary>      
-        [Fact]
-        public async Task ToResultCollectionTaskAsync_ReadOnlyCollection_HasErrors()
-        {
-            var error = CreateErrorTest();
-            var resultHasError = ResultValueFactory.CreateTaskResultValueError<ReadOnlyCollection<int>>(error);
-
-            var resultValue = await resultHasError.ToResultCollectionTaskAsync();
-
-            Assert.True(resultValue.HasErrors);
-            Assert.Single(resultValue.Errors);
-            Assert.True(error.Equals(resultValue.Errors.Last()));
-        }
-
-
-        /// <summary>
-        /// Вернуть результирующий ответ задачи-объекта с коллекцией без ошибок
-        /// </summary>      
-        [Fact]
-        public async Task ToResultCollectionTaskAsync_List_OkStatus()
-        {
-            var collection = GetRangeNumber().ToList();
-            var resultNoError = ResultValueFactory.CreateTaskResultValue(collection);
-
-            var resultValue = await resultNoError.ToResultCollectionTaskAsync();
-
-            Assert.True(resultValue.OkStatus);
-            Assert.True(collection.SequenceEqual(resultValue.Value));
-        }
-
-        /// <summary>
-        /// Вернуть результирующий ответ задачи-объекта со значением с ошибкой
-        /// </summary>      
-        [Fact]
-        public async Task ToResultCollectionTaskAsync_List_HasErrors()
-        {
-            var error = CreateErrorTest();
-            var resultHasError = ResultValueFactory.CreateTaskResultValueError<List<int>>(error);
-
-            var resultValue = await resultHasError.ToResultCollectionTaskAsync();
-
-            Assert.True(resultValue.HasErrors);
-            Assert.Single(resultValue.Errors);
-            Assert.True(error.Equals(resultValue.Errors.Last()));
-        }
-
         /// <summary>
         /// Проверить объект на нул для задачи-объекта. Без ошибок
         /// </summary>
