+++ ResultFunctionalXUnit/FunctionalExtensions/Sync/ResultExtension/ResultCollectionTest/ToResultCollectionExtensionsTest.cs
@@ -1,47 +1,92 @@
-﻿using System.Linq;
+﻿using System.Collections.Generic;
+using System.Linq;
 using ResultFunctional.FunctionalExtensions.Sync.ResultExtension.ResultCollections;
-using ResultFunctional.Models.Interfaces.Results;
-using ResultFunctionalXUnit.Data;
+using ResultFunctional.FunctionalExtensions.Sync.ResultExtension.ResultValues;
+using ResultFunctional.Models.Implementations.Results;
 using Xunit;
 using static ResultFunctionalXUnit.Data.ErrorData;
 using static ResultFunctionalXUnit.Data.Collections;
 
-namespace ResultFunctionalXUnit.FunctionalExtensions.Sync.ResultExtension.ResultCollectionTest
+namespace ResultFunctionalXUnit.FunctionalExtensions.Sync.ResultExtension.ResultCollectionTest;
+
+/// <summary>
+/// Преобразование значений в результирующий ответ коллекции. Тесты
+/// </summary>
+public class ToResultCollectionExtensionsTest
 {
     /// <summary>
-    /// Преобразование значений в результирующий ответ коллекции с условием. Тесты
-    /// </summary>
-    public class ToResultCollectionExtensionsTest
+    /// Вернуть результирующий ответ с коллекцией без ошибок
+    /// </summary>      
+    [Fact]
+    public void ToResultCollection_OkStatus()
     {
-        /// <summary>
-        /// Преобразовать значения в результирующий ответ с условием. Положительное условие
-        /// </summary>
-        [Fact]
-        public void ToResultCollectionWhere_Ok()
-        {
-            var initialCollection = GetRangeNumber();
+        var resultNoError = new ResultError();
+        var collection = new List<string> { "OkStatus" };
 
-            var result = initialCollection.ToResultCollectionWhere(_ => true,
-                                                                   _ => CreateErrorTest());
+        var resultValue = resultNoError.ToResultCollection(collection);
 
-            Assert.True(result.OkStatus);
-            Assert.True(initialCollection.SequenceEqual(result.Value));
-        }
+        Assert.True(resultValue.OkStatus);
+        Assert.True(collection.SequenceEqual(resultValue.Value));
+    }
 
-        /// <summary>
-        /// Преобразовать значения в результирующий ответ с условием. Негативное условие
-        /// </summary>
-        [Fact]
-        public void ToResultCollectionWhere_BadError()
-        {
-            var initialCollection = GetRangeNumber();
-            var errorInitial = CreateErrorTest();
+    /// <summary>
+    /// Вернуть результирующий ответ с коллекцией с ошибкой
+    /// </summary>      
+    [Fact]
+    public void ToResultCollection_HasErrors()
+    {
+        var error = CreateErrorTest();
+        var resultHasError = new ResultError(error);
+        var collection = new List<string> { "BadStatus" };
 
-            var result = initialCollection.ToResultCollectionWhere(_ => false,
-                                                                   _ => errorInitial);
+        var resultValue = resultHasError.ToResultValue(collection);
 
-            Assert.True(result.HasErrors);
-            Assert.True(result.Errors.First().Equals(errorInitial));
-        }
+        Assert.True(resultValue.HasErrors);
+        Assert.Single(resultValue.Errors);
+        Assert.True(error.Equals(resultValue.Errors.Last()));
+    }
+
+    /// <summary>
+    /// Проверить объект на нул. Без ошибок
+    /// </summary>
+    [Fact]
+    public void ToResultCollectionNullCheck_Ok()
+    {
+        var collection = new List<string?> { "1", "2", "3" };
+
+        var resultString = collection.ToResultCollectionNullCheck(CreateErrorTest());
+
+        Assert.True(resultString.OkStatus);
+        Assert.True(collection.SequenceEqual(resultString.Value));
+    }
+
+    /// <summary>
+    /// Проверить объект на нул. Ошибка нулевого значения
+    /// </summary>
+    [Fact]
+    public void ToResultCollectionNullCheck_ErrorCollectionNull()
+    {
+        List<string?>? collection = null;
+        var initialError = CreateErrorTest();
+
+        var resultString = collection.ToResultCollectionNullCheck(initialError);
+
+        Assert.True(resultString.HasErrors);
+        Assert.True(resultString.Errors.First().Equals(initialError));
+    }
+
+    /// <summary>
+    /// Проверить объект на нул. Ошибка нулевого значения
+    /// </summary>
+    [Fact]
+    public void ToResultCollectionNullCheck_ErrorNull()
+    {
+        var collection = new List<string?> { "1", null, "3" };
+        var initialError = CreateErrorTest();
+
+        var resultString = collection.ToResultCollectionNullCheck(initialError);
+
+        Assert.True(resultString.HasErrors);
+        Assert.True(resultString.Errors.First().Equals(initialError));
     }
 }
