+++ ResultFunctionalXUnit/ResultFunctionalXUnit.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <IsPackable>false</IsPackable>
     <Authors>Ivan Rubilnik</Authors>
     <Company>Ivan Rubilnik</Company>
